move Serialize from rspace to shared project
