+++ rholang/src/test/scala/coop/rchain/rholang/interpreter/ReduceSpec.scala
@@ -1,7 +1,6 @@
 package coop.rchain.rholang.interpreter
 
 import cats.implicits._
-
 import com.google.protobuf.ByteString
 import coop.rchain.catscontrib.TaskContrib._
 import coop.rchain.crypto.codec.Base16
@@ -19,11 +18,12 @@ import coop.rchain.rholang.interpreter.errors._
 import coop.rchain.rholang.interpreter.storage._
 import coop.rchain.rspace._
 import coop.rchain.rspace.internal.{Datum, Row, WaitingContinuation}
-
+import coop.rchain.shared.Serialize
 import monix.eval.Task
 import monix.execution.Scheduler.Implicits.global
 import org.scalatest.prop.TableDrivenPropertyChecks._
 import org.scalatest.{AppendedClues, Assertion, FlatSpec, Matchers}
+
 import scala.collection.immutable.BitSet
 import scala.collection.mutable.HashMap
 import scala.collection.{mutable, SortedSet}
