+++ rspace/src/main/scala/coop/rchain/rspace/examples/AddressBookExample.scala
@@ -11,7 +11,7 @@ import coop.rchain.rspace.history.Branch
 import coop.rchain.rspace._
 import coop.rchain.rspace.{RSpace, ReplayRSpace}
 import coop.rchain.shared.Language.ignore
-import coop.rchain.shared.Log
+import coop.rchain.shared.{Log, Serialize}
 import coop.rchain.rspace.util._
 
 import scala.concurrent.ExecutionContext
