+++ rspace/src/test/scala/coop/rchain/rspace/trace/EventTests.scala
@@ -6,7 +6,7 @@ import coop.rchain.rspace.examples.StringExamples.implicits._
 import coop.rchain.rspace.examples.StringExamples.{Pattern, StringsCaptor}
 import coop.rchain.rspace.internal._
 import coop.rchain.rspace.test.ArbitraryInstances._
-import coop.rchain.rspace.{Blake2b256Hash, Serialize}
+import coop.rchain.rspace.Blake2b256Hash
 import org.scalatest.prop.GeneratorDrivenPropertyChecks
 import org.scalatest.{FlatSpec, Matchers}
 import scodec.bits.{BitVector, ByteVector}
@@ -14,6 +14,7 @@ import scodec.codecs.{ignore => cignore, _}
 import scodec.interop.cats._
 import scodec.{Attempt, Codec}
 import coop.rchain.rspace.util
+import coop.rchain.shared.Serialize
 
 import scala.collection.immutable.Seq
